requirements.txt
flask
requests
sentence-transformers
scikit-learn
numpy
python-dotenv
